backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic==2.4.2
numpy==1.24.3
pydantic-settings==2.0.3
mediapipe==0.10.9
tensorflow==2.15.0
scipy==1.11.4
